extract transaction completion logic into own dao
