+++ backend-transaction/src/main/java/de/cotto/bitbook/backend/transaction/TransactionCompletionDao.java
@@ -1,0 +1,7 @@
+package de.cotto.bitbook.backend.transaction;
+
+import java.util.Set;
+
+public interface TransactionCompletionDao {
+    Set<String> getTransactionHashesStartingWith(String hashPrefix);
+}
